v3/requirements.txt
discord.py~=2.4.0
anthropic~=0.40.0
openai>=1.0.0
python-dotenv~=1.0.0
requests~=2.32.0
thefuzz[speedup]~=0.22.0
aiohttp~=3.11.0
